+++ tests/integration/transform-import/tests/index.test.ts
@@ -3,6 +3,8 @@ import { modernBuild } from '../../../utils/modernTestUtils';
 
 const appDir = path.resolve(__dirname, '../');
 
+rstest.setConfig({ testTimeout: 1000 * 60 * 2, hookTimeout: 1000 * 60 * 2 });
+
 describe('test transformImport type', () => {
   let buildRes: { code: number };
   beforeAll(async () => {