+++ packages/cli/plugin-ssg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-ssg
 
+## 3.3.0
+
+### Patch Changes
+
+- @modern-js/prod-server@3.3.0
+- @modern-js/utils@3.3.0
+
 ## 3.2.2
 
 ### Patch Changes